+++ src/meetups/meetups.router.ts
@@ -1,11 +1,11 @@
 import { Router } from "express";
 import { validateRequestProperty } from "../middleware/validateRequestProperty.middleware";
-import { CreateMeetupSchema } from "@schemas/createMeetup.schema";
-import { UpdateMeetupSchema } from "@schemas/updateMeetup.schema";
-import { MeetupIdSchema } from "@schemas/meetupId.schema";
+import { CreateMeetupSchema } from "./schemas/createMeetup.schema";
+import { UpdateMeetupSchema } from "./schemas/updateMeetup.schema";
+import { MeetupIdSchema } from "./schemas/meetupId.schema";
 import { checkRole } from "../middleware/checkRole.middleware";
 import { Roles } from "@authInterfaces/roles.enum";
-import { PaginationQueryParamsSchema } from "@schemas/paginationQueryParams.schema";
+import { PaginationQueryParamsSchema } from "./schemas/paginationQueryParams.schema";
 import { MeetupsController } from "./meetups.controller";
 
 const meetupsRouter = Router();
