+++ src/middleware/sendErrorInCaseOfWrongRoute.middleware.ts
@@ -1,0 +1,5 @@
+import { BadRequestError } from "@responses/httpErrors";
+
+export function sendErrorInCaseOfWrongRoute(req, _, next) {
+  next(BadRequestError(`Cannot find ${req.originalUrl}`));
+}
